docs!: major restructuring of the docs
The docs are now structured in the following structure:

Documentation:
This are the project docs for the developers that is technical.
We follow diataxis.fr guidelines for our documentation structure

Project:
Here are all the moms, sprint reviews and reports located for
the scrum project.

Guide:
This stays the same and continues to be the user facing documentation
for our Franklyn products.